Added module API implementation
