+++ experiments/experiments-impl/build.gradle
@@ -31,6 +31,7 @@ dependencies {
     implementation project(path: ':common-utils')
     implementation project(path: ':experiments-api')
     implementation project(path: ':statistics')
+    implementation project(path: ':backup-agent-api')
     implementation project(path: ':app-build-config-api')
 
     // Room
