+++ experiments/experiments-impl/src/main/java/com/duckduckgo/experiments/impl/VariantManagerImpl.kt
@@ -17,8 +17,8 @@
 package com.duckduckgo.experiments.impl
 
 import androidx.annotation.WorkerThread
-import com.duckduckgo.app.statistics.ReinstallAtbListener.Companion.REINSTALL_VARIANT
 import com.duckduckgo.appbuildconfig.api.AppBuildConfig
+import com.duckduckgo.backup.agent.api.BackupAgentManager
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.experiments.api.VariantConfig
 import com.duckduckgo.experiments.api.VariantManager
@@ -34,6 +34,7 @@ class VariantManagerImpl @Inject constructor(
     private val indexRandomizer: IndexRandomizer,
     private val appBuildConfig: AppBuildConfig,
     private val experimentVariantRepository: ExperimentVariantRepository,
+    private val backupAgentManager: BackupAgentManager,
 ) : VariantManager {
 
     override fun defaultVariantKey(): String {
@@ -63,7 +64,7 @@ class VariantManagerImpl @Inject constructor(
             return
         }
 
-        if (currentVariantKey == REINSTALL_VARIANT) {
+        if (backupAgentManager.isReinstallUser(currentVariantKey)) {
             return
         }
