+++ experiments/experiments-impl/src/test/java/com/duckduckgo/experiments/impl/ExperimentationVariantManagerTest.kt
@@ -17,6 +17,7 @@
 package com.duckduckgo.experiments.impl
 
 import com.duckduckgo.appbuildconfig.api.AppBuildConfig
+import com.duckduckgo.backup.agent.api.BackupAgentManager
 import com.duckduckgo.experiments.api.VariantConfig
 import com.duckduckgo.experiments.impl.store.ExperimentVariantEntity
 import java.util.Locale
@@ -38,6 +39,7 @@ class ExperimentationVariantManagerTest {
     private val appBuildConfig: AppBuildConfig = mock()
     private val activeVariants = mutableListOf<Variant>()
     private val mockExperimentVariantRepository: ExperimentVariantRepository = mock()
+    private val mockBackupAgentManager: BackupAgentManager = mock()
 
     @Before
     fun setup() {
@@ -48,6 +50,7 @@ class ExperimentationVariantManagerTest {
             mockRandomizer,
             appBuildConfig,
             mockExperimentVariantRepository,
+            mockBackupAgentManager,
         )
     }
 
@@ -104,8 +107,10 @@ class ExperimentationVariantManagerTest {
 
     @Test
     fun givenReturnUserVariantWhenVariantsConfigUpdatedThenNewVariantNoAllocated() {
+        val reinstallVariant = "reinstall"
         val variantsConfig = listOf(VariantConfig("variant1", 1.0), VariantConfig("variant2", 1.0))
-        whenever(mockExperimentVariantRepository.getUserVariant()).thenReturn("ru")
+        whenever(mockExperimentVariantRepository.getUserVariant()).thenReturn(reinstallVariant)
+        whenever(mockBackupAgentManager.isReinstallUser(reinstallVariant)).thenReturn(true)
 
         testee.saveVariants(variantsConfig)
